Improve solution 212 to use line-sweep algorithm
Runtime is now under 15 seconds, which is acceptable.
Leaves in the previous option and an additional one as compile time
options.
The third option is similar to the original, but considers all cuboids
at once rather than slicing along the z-axis. It performs similarly to
the original (slightly worse for 40k-50k cuboids) but performs much
better for smaller numbers of cuboids.